refactor(core): extract diff helpers from runners
