+++ crates/flutterdec-core/src/lib.rs
@@ -233,6 +233,7 @@ include!("pipeline/helpers.rs");
 include!("pipeline/model.rs");
 include!("pipeline/quality.rs");
 include!("pipeline/runners_scripts.rs");
+include!("pipeline/runners_diff.rs");
 include!("pipeline/runners.rs");
 include!("pipeline/symbol_map.rs");
 include!("pipeline/engine_fingerprint.rs");
